requirements.txt
mysql-connector-python
google-generativeai
python-dotenv
flask
python-dotenv
pandas
packaging
flask-sqlalchemy
#-e .